Allocate separate JSON lines for each emoji
Achieved by replacing:
`},{`
with
`},\n{`
and
`,"emoticon":`
with
`,\n "emoticon":`

There can still be room for further whitespace correctness such as
is demonstrated by jq output or in the OpenStack JSON conventions.

In partially keeping with the spirit of minified file size of the
original database, however, for now only three bytes per emote is
being contributed (a line feed byte between `},` and `{` and two
more bytes in the form of " \n" between `,` and `"emoticon"`).

The other rationale for this middleground is to keep the file
self-documenting.  (Line numbers are indexible to find records
in a database too big to try to capture in a help manual file.)

The fixed properties (at least up until "tags") that the user
doesn't generally work with get one line; the unpredictable-width
property text defining methods of user input get each second line.

At time of commit, there are 1905 total emoji records here.
Now there are 3812 total lines.  So subtract the two lines for the
opening and closing [,] across the whole file, then divide by two.